代码完善

+++ itmk-base-web/src/main/java/com/itmk/web/sys_menu/controller/SysMenuController.java
@@ -4,19 +4,28 @@ import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.itmk.util.ResultUtils;
 import com.itmk.util.ResultVo;
 import com.itmk.web.sys_menu.entity.MakeMenuTree;
+import com.itmk.web.sys_menu.entity.RouterVO;
 import com.itmk.web.sys_menu.entity.SysMenu;
 import com.itmk.web.sys_menu.service.SysMenuService;
+import com.itmk.web.sys_user.entity.SysUser;
+import com.itmk.web.sys_user.service.SysUserService;
+import org.apache.commons.lang.StringUtils;
 import org.springframework.web.bind.annotation.*;
 
 import javax.annotation.Resource;
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/api/sysMenu")
 public class SysMenuController {
     @Resource
     private SysMenuService sysMenuService;
+    @Resource
+    private SysUserService sysUserService;
 
     //新增
     @PostMapping
@@ -41,6 +50,14 @@ public class SysMenuController {
     //删除
     @DeleteMapping("/{menuId}")
     public ResultVo delete(@PathVariable("menuId") Long menuId){
+        //如果存在下级，不能删除
+        LambdaQueryWrapper<SysMenu> wrapper = new LambdaQueryWrapper<>();
+        wrapper.eq(SysMenu::getParentId,menuId);
+        List<SysMenu> list = sysMenuService.list(wrapper);
+        if (list.size() > 0) {
+            return ResultUtils.error("存在下级，不能删除");
+        }
+
         if (sysMenuService.removeById(menuId)) {
             return ResultUtils.success("删除成功");
         }
@@ -65,6 +82,27 @@ public class SysMenuController {
     public ResultVo getParent(){
         List<SysMenu> list = sysMenuService.getParent();
         return ResultUtils.success("查询成功",list);
+    }
+
+    //获取菜单
+    @GetMapping("/getMenuList")
+    public ResultVo getMenuList(Long userId){
+        //获取用户信息
+        SysUser user = sysUserService.getById(userId);
+        List<SysMenu> menuList = null;
+        //判断是否为超级管理员
+        if(StringUtils.isNotEmpty(user.getIsAdmin()) && "1".equals(user.getIsAdmin())){
+            menuList = sysMenuService.list();
+        }else {
+            menuList = sysMenuService.getMenuByUserId(userId);
+        }
+        //过滤菜单数据,去掉按钮数据
+        Optional.ofNullable(menuList).orElse(new ArrayList<>())
+                .stream()
+                .filter(item -> StringUtils.isNotEmpty(item.getType()) && !item.getType().equals("2")).collect(Collectors.toList());
+        //组装路由数据
+        List<RouterVO> routerList = MakeMenuTree.makeRouter(menuList, 0L);
+        return ResultUtils.success("查询成功",routerList);
     }
 }
